Better version of last commit.
Only grab URI entry focus if the shown webview's URI is the empty
string.  Fixes cases where non-blank new tabs are created and the URI
entry still received focus.

Ideally this would work when switching to a blank notebook tab, but
doing so (at least with the mouse) is moving focus to something else.