Prefer destroying the Window over closing it.
Closing the window (with gtk_window_close) only requests that the
window be closed, presumably by adding a gtk_widget_destroy call to
run later by the main event loop.  This can cause memory corruption
when a fury of ^W keybindings are all sent at once, since it will try
to remove the 0th item from the tabs vector when the vector is already
empty.

Instead, use the destroy_ wrapper method to immediately close the
toplevel window.  I don't see a reason to prefer close to destroy_
from the other occurrences, so change those as well.